src/requirements.txt
opencv-python ~= 4.6.0.66
pyserial ~= 3.5
numpy ~= 1.23.4
pandas ~= 1.5.2
plotly ~= 5.11.0
scikit-learn ~= 1.1.3
scipy ~= 1.9.1
matplotlib ~= 3.6.1
pyqtgraph ~= 0.13.1
requests~=2.27.1